feat: move hackish methods to CashCtrlLedgerExtended class and adapt test suits
